rename generate url
